Added more game logic to the client for testing. Need to prepare game logic on server in proper format before continuing with this. Then need to figure out how to share the game logic project in the server with the unity client so we don't have to mantain code twice!
For client, we are currently checking the state of the game world every frame in order to determine if the player should be able to perform an action. This is very silly! We need to implement a mechanism for "game world state changed" so we can check these sort of things only when we need to.